Add JavaDocs and Follow Coding Standard
