--- data/Hue.txt
@@ -1,2 +1,0 @@
-T | 0 | test
-D | 0 | return book | 2/10/2010 0000
