--- src/main/java/TaskList.java
@@ -1,33 +1,0 @@
-import java.util.ArrayList;
-
-public class TaskList {
-    private final ArrayList<Task> tasks;
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public TaskList(ArrayList<Task> tasks) {
-        this.tasks = tasks;
-    }
-
-    public void add(Task task) {
-        tasks.add(task);
-    }
-
-    public Task get(int index) {
-        return tasks.get(index);
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-
-    public Task remove(int index) {
-        return tasks.remove(index);
-    }
-
-    public ArrayList<Task> getTasks() {
-        return this.tasks;
-    }
-}
