+++ src/main/java/hue/HueException.java
@@ -1,3 +1,4 @@
+package hue;
 public class HueException extends Exception {
     public HueException(String message) {
         super(message);
