+++ src/main/java/hue/command/DeleteCommand.java
@@ -1,0 +1,46 @@
+package hue.command;
+
+import hue.HueException;
+import hue.UI.UI;
+import hue.storage.Storage;
+import hue.task.Task;
+import hue.task.TaskList;
+
+import java.io.IOException;
+
+
+/**
+ * Represents a command to delete a task.
+ */
+public class DeleteCommand extends Command {
+    private final int taskIndex;
+
+    /**
+     * Creates a {@code DeleteCommand} with the specified command string.
+     *
+     * @param fullCommand The full command string.
+     * @throws HueException If the command string is invalid or incomplete.
+     */
+    public DeleteCommand(String fullCommand) throws HueException {
+        try {
+            this.taskIndex = Integer.parseInt(fullCommand.split(" ")[1]) - 1;
+        } catch (NumberFormatException | ArrayIndexOutOfBoundsException e) {
+            throw new HueException("Please provide a valid task number to delete");
+        }
+    }
+
+    @Override
+    public void execute(TaskList tasks, UI ui, Storage storage) throws IOException, HueException {
+        if (taskIndex >= 0 && taskIndex < tasks.size()) {
+            Task removedTask = tasks.remove(taskIndex);
+            ui.showLine();
+            System.out.println("Noted. I've removed this task:");
+            System.out.println("  " + removedTask);
+            System.out.println("Now you have " + tasks.size() + " tasks in the list.");
+
+            storage.saveTasks(tasks);
+        } else {
+            throw new HueException("Task number is out of range.");
+        }
+    }
+}
