+++ src/main/java/hue/command/ExitCommand.java
@@ -1,0 +1,20 @@
+package hue.command;
+
+import hue.UI.UI;
+import hue.storage.Storage;
+import hue.task.TaskList;
+
+/**
+ * Represents a command to exit the application.
+ */
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, UI ui, Storage storage) {
+        ui.showGoodbye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
