+++ src/main/java/hue/task/Task.java
@@ -1,11 +1,24 @@
+package hue.task;
+/**
+ * Represents a generic task with a description and a completion status.
+ */
 public class Task {
     protected String description;
     protected boolean isDone;
-
+    /**
+     * Creates a {@code Task} with the specified description.
+     *
+     * @param description The description of the task.
+     */
     public Task(String description) {
         this.description = description;
         this.isDone = false;
     }
+    /**
+     * Gets the status icon of the task, indicating whether it is done or not.
+     *
+     * @return A string representing the status icon ("X" for done, " " for not done).
+     */
     public String getStatusIcon() {
         return (isDone ? "X" : " ");
     }
@@ -14,15 +27,23 @@ public class Task {
     public String toString() {
         return "[" + getStatusIcon() + "] " + description ;
     }
-
+    /**
+     * Marks the task as done.
+     */
     public void markDone() {
         this.isDone = true;
     }
-
+    /**
+     * Marks the task as not done.
+     */
     public void unmarkDone() {
         this.isDone = false;
     }
-
+    /**
+     * Converts the task to a format suitable for saving to a file.
+     *
+     * @return A string representing the task in file format.
+     */
     public String toFileFormat() {
         String task;
         if (this instanceof Todo) {
