+++ src/main/java/hue/task/Todo.java
@@ -1,4 +1,14 @@
+package hue.task;
+/**
+ * Represents a todo task.
+ */
 public class Todo extends Task {
+
+    /**
+     * Creates a {@code Todo} task with the given description.
+     *
+     * @param description The description of the todo task.
+     */
     public Todo(String description) {
         super(description);
     }
